Support android build
